fix(structure)!: move iot bump commands under the bump command
